Refactor plotting logic to separate module
